Switch from 8-bit to 16-bit Ident value

+++ cuxinterface.cpp
@@ -241,7 +241,7 @@ bool CUXInterface::connectToECU()
 
     uint16_t tune = 0;
     uint8_t checksumFixer = 0;
-    uint8_t ident = 0;
+    uint16_t ident = 0;
 
     if (status)
     {
